Added registration functionality + minor changes to the login fields
