+++ Base Project/.idea/libraries/Maven__com_dlsc_formsfx_formsfx_core_11_3_2.xml
@@ -1,0 +1,13 @@
+<component name="libraryTable">
+  <library name="Maven: com.dlsc.formsfx:formsfx-core:11.3.2">
+    <CLASSES>
+      <root url="jar://$MAVEN_REPOSITORY$/com/dlsc/formsfx/formsfx-core/11.3.2/formsfx-core-11.3.2.jar!/" />
+    </CLASSES>
+    <JAVADOC>
+      <root url="jar://$MAVEN_REPOSITORY$/com/dlsc/formsfx/formsfx-core/11.3.2/formsfx-core-11.3.2-javadoc.jar!/" />
+    </JAVADOC>
+    <SOURCES>
+      <root url="jar://$MAVEN_REPOSITORY$/com/dlsc/formsfx/formsfx-core/11.3.2/formsfx-core-11.3.2-sources.jar!/" />
+    </SOURCES>
+  </library>
+</component>
