+++ Base Project/.idea/libraries/Maven__eu_hansolo_tilesfx_11_48.xml
@@ -1,0 +1,13 @@
+<component name="libraryTable">
+  <library name="Maven: eu.hansolo:tilesfx:11.48">
+    <CLASSES>
+      <root url="jar://$MAVEN_REPOSITORY$/eu/hansolo/tilesfx/11.48/tilesfx-11.48.jar!/" />
+    </CLASSES>
+    <JAVADOC>
+      <root url="jar://$MAVEN_REPOSITORY$/eu/hansolo/tilesfx/11.48/tilesfx-11.48-javadoc.jar!/" />
+    </JAVADOC>
+    <SOURCES>
+      <root url="jar://$MAVEN_REPOSITORY$/eu/hansolo/tilesfx/11.48/tilesfx-11.48-sources.jar!/" />
+    </SOURCES>
+  </library>
+</component>
