+++ Base Project/.idea/libraries/Maven__net_synedra_validatorfx_0_1_13.xml
@@ -1,0 +1,13 @@
+<component name="libraryTable">
+  <library name="Maven: net.synedra:validatorfx:0.1.13">
+    <CLASSES>
+      <root url="jar://$MAVEN_REPOSITORY$/net/synedra/validatorfx/0.1.13/validatorfx-0.1.13.jar!/" />
+    </CLASSES>
+    <JAVADOC>
+      <root url="jar://$MAVEN_REPOSITORY$/net/synedra/validatorfx/0.1.13/validatorfx-0.1.13-javadoc.jar!/" />
+    </JAVADOC>
+    <SOURCES>
+      <root url="jar://$MAVEN_REPOSITORY$/net/synedra/validatorfx/0.1.13/validatorfx-0.1.13-sources.jar!/" />
+    </SOURCES>
+  </library>
+</component>
