+++ Base Project/.idea/libraries/Maven__org_kordamp_ikonli_ikonli_javafx_12_2_0.xml
@@ -1,0 +1,13 @@
+<component name="libraryTable">
+  <library name="Maven: org.kordamp.ikonli:ikonli-javafx:12.2.0">
+    <CLASSES>
+      <root url="jar://$MAVEN_REPOSITORY$/org/kordamp/ikonli/ikonli-javafx/12.2.0/ikonli-javafx-12.2.0.jar!/" />
+    </CLASSES>
+    <JAVADOC>
+      <root url="jar://$MAVEN_REPOSITORY$/org/kordamp/ikonli/ikonli-javafx/12.2.0/ikonli-javafx-12.2.0-javadoc.jar!/" />
+    </JAVADOC>
+    <SOURCES>
+      <root url="jar://$MAVEN_REPOSITORY$/org/kordamp/ikonli/ikonli-javafx/12.2.0/ikonli-javafx-12.2.0-sources.jar!/" />
+    </SOURCES>
+  </library>
+</component>
