+++ Base Project/.idea/modules.xml
@@ -1,0 +1,9 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<project version="4">
+  <component name="ProjectModuleManager">
+    <modules>
+      <module fileurl="file://$PROJECT_DIR$/.idea/Base Project.iml" filepath="$PROJECT_DIR$/.idea/Base Project.iml" />
+      <module fileurl="file://$PROJECT_DIR$/Base Project/Base-Project.iml" filepath="$PROJECT_DIR$/Base Project/Base-Project.iml" />
+    </modules>
+  </component>
+</project>
