+++ Base Project/Base Project/sqlFile1.sql
@@ -1,2 +1,2 @@
 SELECT * FROM user_account;
-INSERT INTO user_account (username, password, balance) VALUES ('user2' , 'parola2' , 300);
+INSERT INTO user_account (username, password, balance) VALUES ('da' , 'das' , 0);
