+++ Base Project/Base Project/src/main/resources/com/example/baseproject/login.fxml
@@ -2,11 +2,14 @@
 
 <?import javafx.scene.control.Button?>
 <?import javafx.scene.control.Label?>
+<?import javafx.scene.control.PasswordField?>
 <?import javafx.scene.control.TextField?>
+<?import javafx.scene.effect.DropShadow?>
 <?import javafx.scene.image.Image?>
 <?import javafx.scene.image.ImageView?>
 <?import javafx.scene.layout.AnchorPane?>
 <?import javafx.scene.layout.BorderPane?>
+<?import javafx.scene.paint.Color?>
 <?import javafx.scene.text.Font?>
 
 <BorderPane maxHeight="-Infinity" maxWidth="-Infinity" minHeight="-Infinity" minWidth="-Infinity" prefHeight="400.0" prefWidth="520.0" xmlns="http://javafx.com/javafx/18" xmlns:fx="http://javafx.com/fxml/1" fx:controller="com.example.baseproject.LoginController">
@@ -24,7 +27,6 @@
     <right>
         <AnchorPane prefHeight="400.0" prefWidth="332.0" style="-fx-background-color: #ECECE7; -fx-text-fill: #FFFFFF;" BorderPane.alignment="CENTER">
             <children>
-                <TextField fx:id="passField" layoutX="76.0" layoutY="109.0" prefHeight="27.0" prefWidth="175.0" promptText="Enter password..." />
                 <Label layoutX="14.0" layoutY="71.0" text="Username" />
                 <Label layoutX="15.0" layoutY="113.0" text="Password" />
                 <ImageView fx:id="lockImageView" fitHeight="34.0" fitWidth="19.0" layoutX="134.0" layoutY="14.0" pickOnBounds="true" preserveRatio="true">
@@ -32,25 +34,33 @@
                         <Image url="@Lock.png" />
                     </image>
                 </ImageView>
-                <Button fx:id="loginButton" layoutX="76.0" layoutY="148.0" mnemonicParsing="false" onAction="#loginButtonOnAction" prefHeight="29.0" prefWidth="56.0" style="-fx-background-color: #323D42; -fx-text-fill: #FFFFFF;" text="Login">
+                <Button fx:id="loginButton" layoutX="76.0" layoutY="148.0" mnemonicParsing="false" onAction="#loginButtonOnAction" prefHeight="29.0" prefWidth="56.0" style="-fx-background-color: #323D42; -fx-text-fill: #FFFFFF; -fx-cursor: hand;" text="Login">
                     <font>
                         <Font name="Arial" size="12.356435775756836" />
                     </font>
                 </Button>
-                <Button fx:id="registerButton" layoutX="144.0" layoutY="148.0" mnemonicParsing="false" onAction="#registrationButtonOnAction" prefHeight="29.0" prefWidth="88.0" style="-fx-background-color: #323D42; -fx-text-fill: #FFFFFF;" text="Register">
+                <Button fx:id="registerButton" layoutX="144.0" layoutY="148.0" mnemonicParsing="false" prefHeight="29.0" prefWidth="88.0" style="-fx-background-color: #323D42; -fx-text-fill: #FFFFFF; -fx-cursor: hand;" text="Register">
                     <font>
                         <Font name="Arial" size="12.356435775756836" />
                     </font>
                 </Button>
-                <Button fx:id="logAdmButton" alignment="CENTER" layoutX="25.0" layoutY="356.0" mnemonicParsing="false" prefHeight="29.0" prefWidth="127.0" style="-fx-background-color: #323D42; -fx-text-fill: #FFFFFF;" text="Login as Admin">
+                <Button fx:id="logAdmButton" alignment="CENTER" layoutX="25.0" layoutY="356.0" mnemonicParsing="false" prefHeight="29.0" prefWidth="127.0" style="-fx-background-color: #323D42; -fx-text-fill: #FFFFFF; -fx-cursor: hand;" text="Login as Admin">
                     <font>
                         <Font name="Arial" size="12.356435775756836" />
                     </font>
                 </Button>
-            <Button fx:id="cancelButton" layoutX="232.0" layoutY="1.0" mnemonicParsing="false" onAction="#cancelButtonOnAction" style="-fx-background-color: #FF0000;" text="Cancel" textFill="WHITE" />
-                <TextField fx:id="userField" layoutX="76.0" layoutY="67.0" prefHeight="27.0" prefWidth="175.0" promptText="Enter username..." />
-            <Label fx:id="invalidLogin" layoutX="55.0" layoutY="209.0" textFill="#fe0303" />
-            <Label fx:id="newRegistration" layoutX="82.0" layoutY="228.0" prefHeight="19.0" prefWidth="162.0" textFill="RED" />
+            <Button fx:id="cancelButton" layoutX="232.0" layoutY="1.0" mnemonicParsing="false" onAction="#cancelButtonOnAction" style="-fx-background-color: #FF0000; -fx-cursor: hand;" text="Cancel" textFill="WHITE" />
+                <TextField fx:id="userField" layoutX="76.0" layoutY="67.0" prefHeight="27.0" prefWidth="175.0" promptText="Enter username..." style="-fx-border-color: #a6a6a6;" />
+            <Label fx:id="invalidLogin" layoutX="55.0" layoutY="209.0" prefHeight="19.0" prefWidth="203.0" textFill="RED" />
+            <Label fx:id="newRegistration" layoutX="19.0" layoutY="228.0" prefHeight="19.0" prefWidth="226.0" textFill="RED">
+               <effect>
+                  <DropShadow>
+                     <color>
+                        <Color red="1.0" green="0.030735455453395844" blue="0.030735455453395844" />
+                     </color>
+                  </DropShadow>
+               </effect></Label>
+            <PasswordField fx:id="passField" layoutX="76.0" layoutY="109.0" prefHeight="28.0" prefWidth="176.0" promptText="Enter password..." style="-fx-border-color: #a6a6a6;" />
             </children>
         </AnchorPane>
     </right>
